backend/pyproject.toml
[project]
name = "chattbc-backend"
version = "0.1.0"
description = "chattbc.gg backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "pydantic>=2,<3",
    "pydantic-settings>=2,<3",
    "sqlalchemy[asyncio]>=2,<3",
    "asyncpg>=0.30,<1",
    "alembic>=1.14,<2",
    "bcrypt>=4,<5",
    "pyjwt>=2,<3",
    "httpx>=0.28,<1",
    "redis>=5,<6",
    "structlog>=24,<26",
    "email-validator>=2,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.25,<1",
    "httpx>=0.28,<1",
    "ruff>=0.11,<1",
    "mypy>=1.14,<2",
]

[tool.ruff]
target-version = "py312"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "S", "T20", "SIM", "RUF"]
ignore = ["S105", "S106", "B008"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]
"alembic/**/*.py" = ["I001", "E501", "W291"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
